refactor: make post as a component
